Logic to populate sales_design table from json received from client side
