+++ salesapp-microservices/salesapp/src/main/java/com/att/salesexpress/webapp/pojos/ResiliencyDesignDO.java
@@ -1,0 +1,23 @@
+package com.att.salesexpress.webapp.pojos;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class ResiliencyDesignDO {
+	@JsonProperty("selectResiliencyOption")
+	private String resiliencyOption;
+
+	public String getResiliencyOption() {
+		return resiliencyOption;
+	}
+
+	public void setResiliencyOption(String resiliencyOption) {
+		this.resiliencyOption = resiliencyOption;
+	}
+
+	@Override
+	public String toString() {
+		return "ResiliencyDesignDO [resiliencyOption=" + resiliencyOption + "]";
+	}
+}
+
+
